# openapi.yaml
openapi: 3.0.3
info:
  title: Fly-Fleet API
  description: Private Jet Charter Brokerage Platform API
  version: 1.0.0
  contact:
    name: Fly-Fleet Support
    email: contact@fly-fleet.com
  license:
    name: Private
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/analytics/internal:
    get:
      summary: List analytics
      description: Track user events and analytics
      tags:
        - analytics
      responses:
        '200':
          description: Success
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create analytics
      description: Track user events and analytics
      tags:
        - analytics
      responses:
        '200':
          description: Success
        '500':
          $ref: '#/components/responses/InternalError'
  /api/contact:
    post:
      summary: Create contact
      description: Handle contact form submissions and inquiries
      tags:
        - contact
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactFormSchema'
  /api/content/{page}/{locale}:
    get:
      summary: List content
      description: Manage CMS content with multilingual support
      tags:
        - content
      responses:
        '200':
          description: Success
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
        - name: page
          in: path
          required: true
          schema:
            type: string
          description: Page parameter
        - name: locale
          in: path
          required: true
          schema:
            type: string
          description: Locale parameter
...
  /api/faqs/{locale}:
    get:
      summary: List faqs
      description: Manage frequently asked questions
      tags:
        - faqs
      responses:
        '200':
          description: Success
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
          description: Maximum number of results
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: search
          in: query
          schema:
            type: string
          description: Search term
        - name: grouped
          in: query
          schema:
            type: boolean
          description: Group results
        - name: locale
          in: path
          required: true
          schema:
            type: string
          description: Locale parameter
...
components:
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin authentication using Bearer token with x-admin-email header
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Validation failed
              details:
                type: array
                items:
                  type: object
                  properties:
                    code:
                      type: string
                    path:
                      type: array
                      items:
                        oneOf:
                          - type: string
                          - type: integer
                    message:
                      type: string
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Rate limit exceeded. Maximum 7 requests per hour.
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Internal server error